@@ -2,8 +2,8 @@
     "name": "rmcgirr83/activity24hours",
     "type": "phpbb-extension",
     "description": "Extension lists of all registered users who have visited the board in the last 24 hours as well as lists stats on the number of new posts, new topics and new users within the last 24 hours",
-    "version": "1.1.2",
-    "time": "2020-07-13",
+    "version": "1.2.0",
+    "time": "2020-07-20",
     "keywords": [
         "phpbb",
         "extension",
@@ -20,7 +20,7 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3",
+        "php": ">=7.1",
         "phpbb/phpbb": ">=3.2",
         "composer/installers": "~1.0.0"
     },
